+++ cocos/3d/CCMeshSkin.cpp
@@ -86,9 +86,9 @@ Bone3D* MeshSkin::getBoneByName(const std::string& id) const
 
 int MeshSkin::getBoneIndex(Bone3D* bone) const
 {
-    for (int i = 0, size = _skinBones.size(); i < size; ++i) {
+    for (size_t i = 0, size = _skinBones.size(); i < size; ++i) {
         if (_skinBones.at(i) == bone)
-            return i;
+            return static_cast<int>(i);
     }
 
     return -1;
